requirements.txt
flask
python-dotenv
scikit-learn
pandas
numpy
